+++ cider-repl.el
@@ -706,8 +706,8 @@ text property `cider-old-input'."
 If invoked in a REPL buffer the command will prompt you for the name of the
 namespace to switch to."
   (interactive (list (if (derived-mode-p 'cider-repl-mode)
-                         (completing-read "Switch to namespace: "
-                                          (cider--all-ns))
+                         (cider-completing-read "Switch to namespace: "
+                                                (cider--all-ns))
                        (cider-current-ns))))
   (if ns
       (with-current-buffer (cider-current-repl-buffer)
@@ -982,8 +982,8 @@ constructs."
   (interactive)
   (if (> (point) cider-repl-input-start-mark)
       (insert (string cider-repl-shortcut-dispatch-char))
-    (let ((command (completing-read "Command: "
-                                   (cider-repl--available-shortcuts))))
+    (let ((command (cider-completing-read "Command: "
+                                          (cider-repl--available-shortcuts))))
      (if (not (equal command ""))
          (call-interactively (gethash command cider-repl-shortcuts))
        (error "No command selected")))))