+++ CHANGELOG.md
@@ -19,6 +19,7 @@ cider-nrepl's info middleware for jump-to-definition.
 * New inspector inspired by SLIME's inspector
 * STDERR ouput is now font-locked with `cider-repl-err-output-face` to make it
 visually distinctive from `cider-repl-output-face` (used for STDOUT output).
+* New `defcustom` `cider-completion-system` controls the completion system employed by CIDER.
 
 ### Changes